requirements.txt
Django
django-browser-reload
django-compressor
django-debug-toolbar
django-filter
django-pwa
django-rosetta
django-tailwind
djangorestframework
django-cors-headers
django-oauth-toolkit
gunicorn
Markdown
numpy
pandas
pika
Pillow
pipenv
psycopg2-binary
python-dotenv
python-gettext
pytz
PyYAML
requests
django-redis
pygments